fix up some warnings in the Query hierarchy
